Move placeholder to its own file
